+++ template/cordova/lib/run.js
@@ -32,10 +32,10 @@ module.exports.run = function (argv) {
         return Q.reject('Could not find project at ' + ROOT);
     }
 
-    // parse args
+    // parse arg
     var args  = nopt({'debug': Boolean, 'release': Boolean, 'nobuild': Boolean,
         'device': Boolean, 'emulator': Boolean, 'target': String, 'archs': String,
-        'phone': Boolean, 'win': Boolean}, {'r' : '--release'}, argv);
+        'phone': Boolean, 'win': Boolean, 'appx': String}, {'r' : '--release'}, argv);
 
     // Validate args
     if (args.debug && args.release) {
@@ -52,11 +52,43 @@ module.exports.run = function (argv) {
     var buildType    = args.release ? 'release' : 'debug',
         buildArchs   = args.archs ? args.archs.split(' ') : ['anycpu'],
         projectType  = args.phone ? 'phone' : 'windows',
-        deployTarget = args.target ? args.target : args.device ? 'device' : 'emulator';
+        deployTarget = args.target ? args.target : (args.emulator ? 'emulator' : 'device'),
+        projOverride = args.appx;
 
     // for win switch we should correctly handle 8.0 and 8.1 version as per configuration
-    if (projectType == 'windows' && getWindowsTargetVersion() == '8.0') {
-        projectType = 'windows80';
+    // as well as 10
+    var targetWindowsVersion = getWindowsTargetVersion();
+    if (projectType == 'windows') {
+        if (targetWindowsVersion == '8.0') {
+            projectType = 'windows80';
+        }
+        else if (targetWindowsVersion === '10.0') {
+            projectType = 'windows10';
+        }
+    }
+    else if (projectType === 'phone') {
+        if (targetWindowsVersion === '10.0' || targetWindowsVersion === 'UAP') {
+            projectType = 'windows10';
+        }
+    }
+
+    if (projOverride) {
+        switch (projOverride) {
+            case '8.1-phone':
+                projectType = 'phone';
+                targetWindowsVersion = '8.1';
+                break;
+            case '8.1-win':
+                projectType = 'windows';
+                targetWindowsVersion = '8.1';
+                break;
+            case 'uap':
+                projectType = 'windows10';
+                break;
+            default:
+                console.warn('Unrecognized --appx parameter passed to run: "' + projOverride + '", ignoring.');
+                break;
+        }
     }
 
     // if --nobuild isn't specified then build app first
@@ -65,10 +97,35 @@ module.exports.run = function (argv) {
     return buildPackages.then(function () {
         return packages.getPackage(projectType, buildType, buildArchs[0]);
     }).then(function(pkg) {
-        console.log('\nDeploying ' + pkg.type + ' package to ' + deployTarget + ':\n' + pkg.file);
-        return pkg.type == 'phone' ?
-            packages.deployToPhone(pkg.file, deployTarget) :
-            packages.deployToDesktop(pkg.file, deployTarget);
+        console.log('\nDeploying ' + pkg.type + ' package to ' + deployTarget + ':\n' + pkg.appx);
+        switch (pkg.type) {
+            case 'phone':
+                return packages.deployToPhoneAndRun(pkg, deployTarget, false).catch(function(e) {
+                    if (args.target || args.emulator || args.device) {
+                        throw e; // Explicit target, carry on
+                    }
+                    
+                    // 'device' was inferred initially, because no target was specified
+                    return packages.deployToPhoneAndRun(pkg, 'emulator', false);
+                });
+                
+            case 'windows10':
+                if (args.phone) {
+                    // Win10 emulator launch is not currently supported, always force device
+                    if (args.emulator || args.target === 'emulator') {
+                        console.warn('Windows 10 Phone emulator is not currently supported, attempting deploy to device.');
+                    }
+                    return packages.deployToPhoneAndRun(pkg, 'device', true);
+                }
+                else {
+                    return packages.deployToDesktop(pkg, deployTarget, projectType);
+                }
+            
+                break;    
+            default: // 'windows'
+                return packages.deployToDesktop(pkg, deployTarget, projectType);
+                
+        }
     });
 };
 
@@ -84,6 +141,9 @@ module.exports.help = function () {
     console.log('    --archs       : Specific chip architectures (`anycpu`, `arm`, `x86`, `x64`).');
     console.log('    --phone, --win');
     console.log('                  : Specifies project type to deploy');
+    console.log('    --appx=<8.1-win|8.1-phone|uap>');
+    console.log('                  : Overrides WindowsTargetVersion to build Windows 8.1, ');
+    console.log('                              Windows Phone 8.1, or Windows 10.');
     console.log('');
     console.log('Examples:');
     console.log('    run');
@@ -92,21 +152,26 @@ module.exports.help = function () {
     console.log('    run --target=7988B8C3-3ADE-488d-BA3E-D052AC9DC710');
     console.log('    run --device --release');
     console.log('    run --emulator --debug');
+    console.log('    run --device --appx=phone-8.1');
     console.log('');
+
     process.exit(0);
 };
 
 
 function getWindowsTargetVersion() {
     var config = new ConfigParser(path.join(ROOT, 'config.xml'));
-    var windowsTargetVersion = config.getPreference('windows-target-version');
+    var windowsTargetVersion = config.getWindowsTargetVersion();
     switch(windowsTargetVersion) {
-    case '8':
-    case '8.0':
-        return '8.0';
-    case '8.1':
-        return '8.1';
-    default:
-        throw new Error('Unsupported windows-target-version value: ' + windowsTargetVersion);
+        case '8':
+        case '8.0':
+            return '8.0';
+        case '8.1':
+            return '8.1';
+        case '10.0':
+        case 'UAP':
+            return '10.0';
+        default:
+            throw new Error('Unsupported WindowsTargetVersion value: ' + windowsTargetVersion);
     }
 }